+++ lib/app/app_module.dart
@@ -44,14 +44,14 @@ abstract class AppModule {
     ...SecurityModuleRoutes().pages,
     ...WebViewModuleRoutes().pages,
     ...AppPurchaseModuleRoutes().pages,
-    ...CheckPointModule.routes,
+    ...CheckPointsModule().pages,
   ];
 
   static void setupHomePages() {
     HomePage.initialIndex = 0;
     HomePage.navigatorItems = [
       NavigatorItem(
-        routeName: AppRouter.checkPoint.name,
+        routeName: AppRouter.checkPoints.name,
         bottomItem: NavigatorBottom(
           title: 'home',
           selectedIcon: const FlutterIcon(