+++ lib/presentation/presentation.dart
@@ -1,1 +1,1 @@
-export 'check_point/check_point.dart';
+export 'check_points/check_points.dart';