+++ packages/core/lib/src/router/app_router.dart
@@ -33,7 +33,7 @@ enum AppRouter {
   deleteAccountConfirmation('/delete_account/confirmation'),
   deleteAccountFinished('/delete_account/delete_account_finished'),
   purchase('/purchase'),
-  checkPoint('/check_point')
+  checkPoints('/check_points')
   ;
 
   const AppRouter(this._name);